% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsize_base.R
\name{popsize_base}
\alias{popsize_base}
\title{Estimate total population size and capture probability using user provided set of models.}
\usage{
popsize_base(
  data,
  K = 2,
  j0,
  k0,
  filterrows = FALSE,
  funcname = c("rangerlogit"),
  nfolds = 5,
  margin = 0.005,
  sl.lib = c("SL.gam", "SL.glm", "SL.glm.interaction", "SL.ranger", "SL.glmnet"),
  Nmin = 500,
  TMLE = TRUE,
  PLUGIN = TRUE,
  ...
)
}
\arguments{
\item{data}{The data frame in capture-recapture format for which total population is to be estimated.
The first K columns are the capture history indicators for the K lists. The remaining columns are covariates in numeric format.}

\item{K}{The number of lists in the data. typically the first \code{K} rows of data.}

\item{j0}{The first list to be used for estimation.}

\item{k0}{The secod list to be used in the estimation.}

\item{filterrows}{A logical value denoting whether to remove all rows with only zeroes.}

\item{funcname}{The vector of estimation function names to obtain the population size.}

\item{nfolds}{The number of folds to be used for cross fitting.}

\item{margin}{The minimum value the estimates can attain to bound them away from zero.}

\item{sl.lib}{algorithm library for SuperLearner. Default library includes "gam", "glm", "glmnet", "glm.interaction", "ranger".}

\item{Nmin}{The cutoff for minimum sample size to perform doubly robust estimation. Otherwise, Petersen estimator is returned.}

\item{TMLE}{The logical value to indicate whether TMLE has to be computed.}

\item{PLUGIN}{The logical value to indicate whether the plug-in estimates is returned.}

\item{...}{Any extra arguments passed into the function.}
}
\value{
A list of estimates containing the following components for each list-pair, model and method (PI = plug-in, DR = doubly-robust, TMLE = targeted maximum likelihood estimate):
\item{result}{  A dataframe of the below estimated quantities.
\itemize{
\item{psi}{  The estimated capture probability.}
\item{sigma}{  The efficiency bound.}
\item{n}{  The estimated population size n.}
\item{sigman}{  The estimated standard deviation of the population size.}
\item{cin.l}{  The estimated lower bound of a 95\% confidence interval of \code{n}.}
\item{cin.u}{  The estimated upper bound of a 95\% confidence interval of \code{n}.}}}
\item{N}{  The number of data points used in the estimation after removing rows with missing data.}
\item{ifvals}{  The estimated influence function values for the observed data.}
\item{nuis}{  The estimated nuisance functions (q12, q1, q2) for each element in funcname.}
\item{nuistmle}{  The estimated nuisance functions (q12, q1, q2) from tmle for each element in funcname.}
\item{idfold}{  The division of the rows into sets (folds) for cross-fitting.}
}
\description{
Estimate total population size and capture probability using user provided set of models.
}
\examples{
data = simuldata(n = 6000, l = 3)$data
psin_estimate = popsize_base(data = data[,1:2])
#this returns the basic plug-in estimate since covariates are absent.

psin_estimate = popsize_base(data = data, funcname = c("gam", "rangerlogit"))
#this returns the plug-in, the bias-corrected and the tmle estimate for the two models
}
\references{
Das, M., & Kennedy, E. H. (2021). Doubly robust capture-recapture methods for estimating population size. arXiv preprint arXiv:2104.14091.
}
