% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_psi_n_controlled_errors.R
\name{popsize_simul}
\alias{popsize_simul}
\title{Estimate the total population size and capture probabilities using perturbed true nuisance functions.}
\usage{
popsize_simul(
  List_matrix,
  n,
  K,
  nfolds = 5,
  omega,
  alpha,
  eps = 0.005,
  iter = 100,
  twolist = TRUE
)
}
\arguments{
\item{List_matrix}{The data frame in capture-recapture format for which total population is to be estimated.
The first K columns are the capture history indicators for the K lists. The remaining columns are covariates in numeric format.}

\item{n}{The true population size. Required to calculate the added error.}

\item{K}{The number of lists in the data. typically the first \code{K} rows of List_matrix.}

\item{nfolds}{The number of folds to be used for cross fitting.}

\item{omega}{The standard deviation from zero of the added error.}

\item{alpha}{The rate of convergence. Takes values in (0, 1].}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}

\item{iter}{An integer denoting the maximum number of iterations allowed for targeted maximum likelihood method.}

\item{twolist}{The logical value of whether targeted maximum likelihood algorithm fits only two modes when K = 2.}
}
\value{
A list of estimates containing the following components:
\item{psi}{  A matrix of the estimated capture probability for each list pair, model and method combination. In the absence of covariates, the column represents the standard plug-in estimate.
The rows represent the list pair which is assumed to be independent conditioned on the covariates.
The columns represent the model and method combinations (PI = plug-in, DR = bias-corrected, TMLE = targeted maximum likelihood estimate)indicated in the columns.}
\item{sigma2}{  A matrix of the efficiency bound \code{sigma^2} in the same format as \code{psi}.}
\item{n}{  A matrix of the estimated population size n in the same format as \code{psi}.}
\item{varn}{  A matrix of the variance for population size estimate in the same format as \code{psi}.}
\item{N}{  The number of data points used in the estimation after removing rows with missing data.}
}
\description{
Estimate the total population size and capture probabilities using perturbed true nuisance functions.
}
\examples{
data = matrix(sample(c(0,1), 2000, replace = TRUE), ncol = 2)
x = matrix(rnorm(nrow(data)*3, 2,1), nrow = nrow(data))

psin_estimate = popsize_simul(List_matrix = data, alpha = 0.25, omega = 1)
#this returns the basic plug-in estimate since covariates are absent.

data = cbind(data, x)
psin_estimate = popsize_simul(List_matrix = data, funcname = c("logit", "sl"), nfolds = 2, twolist = FALSE, eps = 0.005, alpha = 0.25, omega = 1)
#this returns the plug-in, the bias-corrected and the tmle estimate for the two models
}
\references{
Gruber, S., & Van der Laan, M. J. (2011). tmle: An R package for targeted maximum likelihood estimation.

van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2008) Super Learner, Statistical Applications of Genetics and Molecular Biology, 6, article 25.
}
