% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psinhatcond.R
\name{psinhatcond}
\alias{psinhatcond}
\title{Estimate total population size and capture probability using user provided set of models conditioned on an attribute.}
\usage{
psinhatcond(
  List_matrix,
  K = 2,
  funcname = c("logit"),
  condvar,
  nfolds = 5,
  twolist = FALSE,
  eps = 0.005,
  iter = 50,
  sl.lib = c("SL.gam", "SL.glm", "SL.glm.interaction", "SL.ranger", "SL.glmnet")
)
}
\arguments{
\item{List_matrix}{The data frame in capture-recapture format for which total population is to be estimated.
The first K columns are the capture history indicators for the K lists. The remaining columns are covariates in numeric format.}

\item{K}{The number of lists in the data. typically the first \code{K} rows of List_matrix.}

\item{funcname}{The vector of estimation function names to obtain the population size.}

\item{condvar}{The covariate for which conditional estimates are required.}

\item{nfolds}{The number of folds to be used for cross fitting.}

\item{twolist}{The logical value of whether targeted maximum likelihood algorithm fits only two modes when K = 2.}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}

\item{iter}{An integer denoting the maximum number of iterations allowed for targeted maximum likelihood method.}

\item{sl.lib}{algorithm library for SuperLearner. Default library includes "gam", "glm", "glmnet", "glm.interaction", "ranger".}
}
\value{
A list of estimates containing the following components:
\item{psi}{  A dataframe of the estimated capture probability for each list pair, model and method combination. In the absence of covariates, the column represents the standard plug-in estimate.
The \code{listpair} column represents the list pair which is assumed to be independent conditioned on the covariates.
The \code{condvar} column stores the value of the original \code{condvar} attribute in \code{List_matrix} which is conditioned upon.
The columns represent the model and method combinations (PI = plug-in, BC = bias-corrected, TMLE = targeted maximum likelihood estimate)indicated in the columns.}
\item{sigma2}{  A dataframe of the efficiency bound \code{sigma^2} in the same format as \code{psi}.}
\item{n}{  A dataframe of the estimated population size n in the same format as \code{psi}.}
\item{varn}{  A dataframe of the variance for population size estimate in the same format as \code{psi}.}
\item{N}{  The number of data points used in the estimation for each value of \code{condvar}.}
\item{cin.l}{  The daraframe of estimated lower bound of a 95\% confidence interval of \code{n}.}
\item{cin.u}{  The dataframe of estimated upper bound of a 95\% confidence interval of \code{n}.}
}
\description{
Estimate total population size and capture probability using user provided set of models conditioned on an attribute.
}
\examples{
data = matrix(sample(c(0,1), 2000, replace = TRUE), ncol = 2)
x = matrix(rnorm(nrow(data)*3, 2,1), nrow = nrow(data))
ss = sample(1:6, nrow(data), replace = TRUE)

data = cbind(data, x, ss)
psin_estimate = psinhatcond(List_matrix = data, funcname = c("logit", "sl"), condvar = 'ss', nfolds = 2, twolist = FALSE, eps = 0.005)
#this returns the plug-in, the bias-corrected and the tmle estimate for the two models conditioned on column ss
}
\references{
Gruber, S., & Van der Laan, M. J. (2011). tmle: An R package for targeted maximum likelihood estimation.

van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2008) Super Learner, Statistical Applications of Genetics and Molecular Biology, 6, article 25.
}
