% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psinhatgivenq.R
\name{psinhatgivenq}
\alias{psinhatgivenq}
\title{Estimate total population size and capture probability using user provided set of models.}
\usage{
psinhatgivenq(
  List_matrix,
  i = 1,
  j = 2,
  eps = 0.005,
  qhateval,
  q1mat,
  q2mat,
  q12mat,
  idfold,
  TMLE = TRUE,
  ...
)
}
\arguments{
\item{List_matrix}{The data frame in capture-recapture format with two lists for which total population is to be estimated.
The first K columns are the capture history indicators for the K lists. The remaining columns are covariates in numeric format.}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}

\item{qhateval}{A list object of class \code{qhateval} as returned by the \code{qhateval} function.}

\item{q1mat}{A dataframe with capture probabilities for the first list.}

\item{q2mat}{A dataframe with capture probabilities for the second list.}

\item{q12mat}{A dataframe with capture probabilities for both the lists simultaneously.}

\item{idfold}{The fold assignment of each row during estimation.}

\item{TMLE}{The logical value to indicate whether TMLE has to be computed.}
}
\value{
A list of estimates containing the following components:
\item{psi}{  A dataframe of the estimated capture probability for each list pair, model and method combination. In the absence of covariates, the column represents the standard plug-in estimate.
The rows represent the list pair which is assumed to be independent conditioned on the covariates.
The columns represent the model and method combinations (PI = plug-in, DR = doubly-robust, TMLE = targeted maximum likelihood estimate)indicated in the columns.}
\item{sigma2}{  A dataframe of the efficiency bound \code{sigma^2} in the same format as \code{psi}.}
\item{n}{  A dataframe of the estimated population size n in the same format as \code{psi}.}
\item{varn}{  A dataframe of the variance for population size estimate in the same format as \code{psi}.}
\item{N}{  The number of data points used in the estimation after removing rows with missing data.}
\item{ifvals}{  The estimated influence function values for the observed data. Each column corresponds to an element in funcname.}
\item{nuis}{  The estimated nuisance functions (q12, q1, q2) for each element in funcname.}
\item{nuistmle}{  The estimated nuisance functions (q12, q1, q2) from tmle for each element in funcname.}
\item{cin.l}{  The estimated lower bound of a 95\% confidence interval of \code{n}.}
\item{cin.u}{  The estimated upper bound of a 95\% confidence interval of \code{n}.}
}
\description{
Estimate total population size and capture probability using user provided set of models.
}
\examples{
data = simuldata(1000, 1)$List_matrix
qhat_estimate = qhateval(List_matrix = data, funcname = c("logit", "gam"), nfolds = 2, eps = 0.005)
psin_estimate = psinhatgivenq(List_matrix = data, qhateval = qhat_estimate)
}
\references{
Gruber, S., & Van der Laan, M. J. (2011). tmle: An R package for targeted maximum likelihood estimation.

van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2008) Super Learner, Statistical Applications of Genetics and Molecular Biology, 6, article 25.
}
