% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhat_estimation_functions.R
\name{qhat_ranger}
\alias{qhat_ranger}
\title{Estimate marginal and joint distribution of lists j and k using ranger.}
\usage{
qhat_ranger(List.train, List.test, K = 2, j = 1, k = 2, margin = 0.005, ...)
}
\arguments{
\item{List.train}{The training data matrix used to estimate the distibution functions.}

\item{List.test}{The data matrix on which the estimator function is applied.}

\item{K}{The number of lists in the data.}

\item{j}{The first list that is conditionally independent.}

\item{k}{The second list that is conditionally independent.}

\item{margin}{The minimum value the estimates can attain to bound them away from zero.}

\item{...}{Any extra arguments passed into the function.}
}
\value{
A list of the marginal and joint distribution probabilities q_1, q_2 and q_12.
}
\description{
Estimate marginal and joint distribution of lists j and k using ranger.
}
\examples{
\dontrun{
qhat = qhat_ranger(List.train = List.train, List.test = List.test, margin = 0.005)
q1 = qhat$q1
q2 = qhat$q2
q12 = qhat$q12
}
}
\references{
Marvin N. Wright, Andreas Ziegler (2017). ranger: A Fast Implementation of Random Forests for High Dimensional Data in C++ and R. Journal of Statistical Software, 77(1), 1-17. doi:10.18637/jss.v077.i01
}
