% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuldata.R
\name{simuldata}
\alias{simuldata}
\title{A function to reorder the columns of a data table/matrix/data frame and to change factor variables to numeric.}
\usage{
simuldata(n, l, categorical = FALSE, ep = 0, K = 2)
}
\arguments{
\item{n}{The size of the population.}

\item{l}{The number of continuous covariates.}

\item{categorical}{A logical value of whether to include a categorical column.}

\item{ep}{A numeric value to change the list probabilities.}
}
\value{
A list of estimates containing the following components:
\item{List_matrix}{  A dataframe in with two list capture histories and covariates from a population if true size \code{n}.}
\item{psi}{  The empirical capture probability for the set-up used.}
}
\description{
A function to reorder the columns of a data table/matrix/data frame and to change factor variables to numeric.
}
\examples{
data = simuldata(n = 1000, l = 2)$List_matrix
psi = simuldata(n = 10000, l = 2)$psi
}
