% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotci.R
\name{plotci}
\alias{plotci}
\title{Plot estimated confidence interval of total population size.}
\usage{
plotci(object, tsize = 12, ...)
}
\arguments{
\item{object}{An object of class \code{popsize} or \code{popsize_cond}.}

\item{tsize}{The text size for the plots.}

\item{...}{Any extra arguments passed into the function.}
}
\value{
A ggplot object \code{fig} with population size estimates and the 95\% confidence intervals.
}
\description{
Plot estimated confidence interval of total population size.
}
\examples{

data = simuldata(n = 10000, l = 1)$List_matrix_xstar

p = popsize(List_matrix = data, funcname = c("logit", "gam"))
plotci(p)

data = simuldata(n = 10000, l = 1, categorical = TRUE)$List_matrix_xstar
p = popsize_cond(List_matrix = data, condvar = 'catcov')
plotci(p)
}
