% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsize.R
\name{popsize}
\alias{popsize}
\title{Estimate total population size and capture probability using user provided set of models or user provided nuisance estimates.}
\usage{
popsize(
  List_matrix,
  i = 1,
  j = 2,
  eps = 0.005,
  getnuis,
  q1mat,
  q2mat,
  q12mat,
  idfold,
  TMLE = TRUE,
  PLUGIN = TRUE,
  ...
)
}
\arguments{
\item{List_matrix}{The data frame in capture-recapture format with \code{K} lists for which total population is to be estimated.
The first K columns are the capture history indicators for the \code{K} lists. The remaining columns are covariates in numeric format.}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}

\item{getnuis}{A list object of class \code{getnuis} as returned by the \code{getnuis} function.}

\item{q1mat}{A dataframe with capture probabilities for the first list.}

\item{q2mat}{A dataframe with capture probabilities for the second list.}

\item{q12mat}{A dataframe with capture probabilities for both the lists simultaneously.}

\item{idfold}{The fold assignment of each row during estimation.}

\item{TMLE}{The logical value to indicate whether TMLE has to be computed.}

\item{PLUGIN}{The logical value to indicate whether the plug-in estimates is returned.}

\item{funcname}{The vector of estimation function names to obtain the population size.}
}
\value{
A list of estimates containing the following components for each list-pair, model and method (PI = plug-in, DR = doubly-robust, TMLE = targeted maximum likelihood estimate):
\item{result}{  A dataframe of the below estimated quantities.
\itemize{
\item{psi}{  The estimated capture probability.}
\item{sigma}{  The efficiency bound.}
\item{n}{  The estimated population size n.}
\item{sigman}{  The estimated standard deviation of the population size.}
\item{cin.l}{  The estimated lower bound of a 95\% confidence interval of \code{n}.}
\item{cin.u}{  The estimated upper bound of a 95\% confidence interval of \code{n}.}}}
\item{N}{  The number of data points used in the estimation after removing rows with missing data.}
\item{ifvals}{  The estimated influence function values for the observed data.}
\item{nuis}{  The estimated nuisance functions (q12, q1, q2) for each element in funcname.}
\item{nuistmle}{  The estimated nuisance functions (q12, q1, q2) from tmle for each element in funcname.}
}
\description{
Estimate total population size and capture probability using user provided set of models or user provided nuisance estimates.
}
\examples{
data = simuldata(1000, 1)$List_matrix
qhat_estimate = getnuis(List_matrix = data, funcname = c("logit", "gam"), nfolds = 2, eps = 0.005)
psin_estimate = popsize(List_matrix = data, getnuis = qhat_estimate)
}
\references{
Gruber, S., & Van der Laan, M. J. (2011). tmle: An R package for targeted maximum likelihood estimation.

van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2008) Super Learner, Statistical Applications of Genetics and Molecular Biology, 6, article 25.
}
