% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhat_estimation_functions.R
\name{qhat_ranger}
\alias{qhat_ranger}
\title{Estimate marginal and joint distribution of lists i and j using ranger.}
\usage{
qhat_ranger(List.train, List.test, K = 2, i = 1, j = 2, eps = 0.005, ...)
}
\arguments{
\item{List.train}{The training data matrix used to estimate the distibution functions.}

\item{List.test}{The data matrix on which the estimator function is applied.}

\item{K}{The number of lists in the data.}

\item{i}{The first list that is conditionally independent.}

\item{j}{The second list that is conditionally independent.}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}
}
\value{
A list of the marginal and joint distribution probabilities q_1, q_2 and q_12.
}
\description{
Estimate marginal and joint distribution of lists i and j using ranger.
}
\examples{
qhat = qhat_ranger(List.train = List.train, List.test = List.test, K = 3, i = 1, j = 2, eps = 0.005)
q1 = qhat$q1
q2 = qhat$q2
q12 = qhat$q12

}
