% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhat_estimation_functions.R
\name{qhat_sl}
\alias{qhat_sl}
\title{Estimate marginal and joint distribution of lists i and j using super learner.}
\usage{
qhat_sl(List1, List2, K, i, j, eps = 0.005, sl.lib = sl.lib)
}
\arguments{
\item{List1}{The training data matrix used to estimate the distibution functions.}

\item{List2}{The data matrix on which the estimator function is applied.}

\item{K}{The number of lists in the data.}

\item{i}{The first list that is conditionally independent.}

\item{j}{The second list that is conditionally independent.}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}
}
\value{
A list of the marginal and joint distribution probabilities q_1, q_2 and q_12.
}
\description{
Estimate marginal and joint distribution of lists i and j using super learner.
}
\examples{
qhat = qhat_sl(List1 = List1, List2 = List2, K = 3, i = 1, j = 2, eps = 0.005)
q1 = qhat$q1
q2 = qhat$q2
q12 = qhat$q12

}
