% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnle_for_psi.R
\name{tmle}
\alias{tmle}
\title{Returns the targeted maximum likelihood estimates for the nuisance functions}
\usage{
tmle(datmat, iter = 100, eps = 0.005, eps_stop = 1e-05, twolist = FALSE)
}
\arguments{
\item{datmat}{The data frame containing columns \code{yi}, \code{yj}, \code{yij}, \code{q10}, \code{q02} and \code{q12}.}

\item{iter}{An integer denoting the maximum number of iterations allowed for targeted maximum likelihood method. Default value is 100.}

\item{eps}{The minimum value the estimates can attain to bound them away from zero.}

\item{eps_stop}{The minimum value the estimates can attain to bound them away from zero.}

\item{twolist}{The logical value of whether targeted maximum likelihood algorithm fits only two modes when K = 2.}
}
\value{
A list of estimates containing the following components:
\item{error}{  An indicator of whether the algorithm ran and converged. Returns FALSE, if it ran correctly and FALSE otherwise.}
\item{datmat}{  A data frame returning \code{datmat} with the updated estimates for the nuisance functions \code{q10}, \code{q02} and \code{q12}. This is returned only if \code{error} is FALSE.}
}
\description{
Returns the targeted maximum likelihood estimates for the nuisance functions
}
\examples{
data = matrix(sample(c(0,1), 2000, replace = TRUE), ncol = 2)
xqmat = matrix(runif(nrow(data)*3, 0, 1), nrow = nrow(data))
datmat = cbind(data, data[,1]*data[,2], xmat)
colnames(datmat) = c("yi", "yj", "yij", "q10", "q02", "q12")
datmat = as.data.frame(datmat)
result = tmle(datmat, eps = 0.005, eps_stop = 0.00001, twolist = TRUE)
}
\references{
Gruber, S., & Van der Laan, M. J. (2011). tmle: An R package for targeted maximum likelihood estimation.
}
